observability: address external review (CI lint + tests + doc nits)
CI lint failures
- observability/option.go: drop trailing blank line at EOF (gofmt).
- observability/resources.go: preallocate baseAttrs slice (prealloc).
- observability/attributes.go: drop redundant uint64() cast on
  spectypes.OperatorID (which is `type OperatorID = uint64`) — matches
  the existing pattern in ValidatorProposerAttribute.

Tests
- observability/configurator_test.go: lock the InitializeLogger /
  Initialize contracts: InitializeLogger propagation into metrics;
  Initialize works without InitializeLogger (no-op fallback);
  Initialize does NOT replace metrics.logger (verified via sentinel +
  RecordUint64Value error trigger). The third test guards the
  "intentionally does not re-propagate" comment from silent regressions.
- beacon/goclient/proposer_verify_metric_test.go: lock the metric-label
  contract — verifyProposalParent's three branches each fire the
  expected counter carrying ssv.beacon.client. Subtests share one OTel
  meter provider because the global delegating-meter only re-binds
  package-level instruments on the first SetMeterProvider call after
  package init; per-test providers don't compose.

Code / comment edits
- beacon/goclient/observability.go: document the invariant that the
  `clients` slice passed to registerProposalParentBaselines must be
  final at call time (the closure captures the slice header).
- beacon/goclient/proposer.go: route the mismatch log's beacon_client
  field through the new fields.BeaconClient helper.
- observability/log/fields/fields.go: add fields.BeaconClient helper
  and FieldBeaconClient constant, mirroring the ssv.beacon.client
  metric attribute (dotted OTel form vs snake_case log convention).
- cli/operator/node.go: rename metrics_enabled / traces_enabled log
  fields to _configured (they reflect config, not runtime state); gate
  EmitBaselines on cfg.MetricsAPIPort > 0 to skip no-op iteration when
  metrics are disabled.
- observability/metrics/baseline.go: explain the locking asymmetry in
  EmitBaselines — sparse-counters loop holds the lock during Add
  because Int64Counter.Add is a leaf and cannot reenter the registry;
  labeled-baselines path snapshots first because user-provided fns may.
- network/peers/connections/observability.go,
  message/validation/observability.go: replace ambiguous "Sparse" /
  "Possibly sparse" comments with TODO(audit) markers — re-evaluate
  these classifications against production rates. Mis-classifying is
  harmless (Add(0) on a dense counter is a no-op for PromQL) but the
  registry is meant to document intent.

+++ observability/attributes.go
@@ -30,7 +30,7 @@ const (
 func OperatorIDAttribute(id spectypes.OperatorID) attribute.KeyValue {
 	return attribute.KeyValue{
 		Key:   "ssv.operator_id",
-		Value: Uint64AttributeValue(uint64(id)),
+		Value: Uint64AttributeValue(id),
 	}
 }
